Moved footer content to match new file structure
