+++ client/src/pages/sub-pages/Trends/TrendsMap.js
